feat(plugin): restructure files
